+++ src/actions/google.ts
@@ -1,6 +1,8 @@
 import {saveResultForBashWrapper} from "../util/fileUtil";
 import {getLastQuestionFromHistory} from "../util/historyUtil";
 
+const GOOGLE_SEARCH_URL = 'https://www.google.pl/search'
+
 export const openLastQuestionInGoogleFromHistory = () => {
     const question = getLastQuestionFromHistory()
     openInGoogle(question)
@@ -10,8 +12,7 @@ export const openQuestionInGoogle = (question: string) => openInGoogle(question)
 
 const openInGoogle = (question: string) => {
     const params = new URLSearchParams({q: question}).toString()
-    const uri = `https://www.google.pl/search?${params}`
+    const uri = `${GOOGLE_SEARCH_URL}?${params}`
     const command = `open "${uri}" > /dev/null 2>&1`
-    console.log('command', command)
     saveResultForBashWrapper(command);
 }
